refactor(lib): dissolve constants/data/serverUrl/previewSecret
Relocate the four vaguely-named src/lib/ root files to purposeful homes:

- constants.ts (PAGE_TAGS, PageTag)       -> collections/Pages/pageTags.ts
- data.ts (GENDER_OPTIONS)                -> lib/utilities/gender.ts
- serverUrl.ts (getServerUrl)             -> lib/utilities/serverUrl.ts
- previewSecret.ts (header + validator)   -> lib/utilities/previewSecret.ts

PAGE_TAGS is page-domain (single owner is Pages). The other three are
cross-boundary shared utilities. Purely structural; no behavior change.

Refs #444
