+++ src/collections/Pages/Pages.ts
@@ -1,10 +1,10 @@
 import type { CollectionConfig } from 'payload'
 
 import { pageBlocks } from '@/blocks'
+import { PAGE_TAGS } from '@/collections/Pages/pageTags'
 import { slugField } from '@/fields'
 import { APP_REQUIRED_PAGE_FIELDS } from '@/globals/WeMeditateAppConfig/WeMeditateAppConfig'
 import { removeDanglingLexicalReferencesAfterRead } from '@/hooks/lexicalHooks'
-import { PAGE_TAGS } from '@/lib/constants'
 import { fullRichTextEditor } from '@/lib/richEditor'
 
 export const Pages: CollectionConfig = {
