+++ src/payload.config.ts
@@ -14,7 +14,7 @@ import { GetPlatformProxyOptions } from 'wrangler'
 
 import { serverEnv } from '@/lib/env'
 import { buildPayloadLocales, DEFAULT_LOCALE } from '@/lib/locales'
-import { getServerUrl } from '@/lib/serverUrl'
+import { getServerUrl } from '@/lib/utilities/serverUrl'
 import { createWorkerSafeLogger } from '@/lib/workerSafeLogger'
 import { accessPlugin, bypassPermissions, filterAvailableLocales } from '@/plugins/access'
 import { resendAdapter } from '@/plugins/email/resendAdapter'
